+++ examples/chat-svelte/CHANGELOG.md
@@ -1,5 +1,20 @@
 # passkey-svelte
 
+## 0.0.100
+
+### Patch Changes
+
+- Updated dependencies [bdc9aee]
+  - jazz-tools@0.15.11
+
+## 0.0.99
+
+### Patch Changes
+
+- Updated dependencies [9815ec6]
+- Updated dependencies [b4fdab4]
+  - jazz-tools@0.15.10
+
 ## 0.0.98
 
 ### Patch Changes