+++ packages/cojson-storage-sqlite/CHANGELOG.md
@@ -1,5 +1,17 @@
 # cojson-storage-sqlite
 
+## 0.15.11
+
+### Patch Changes
+
+- cojson@0.15.11
+
+## 0.15.10
+
+### Patch Changes
+
+- cojson@0.15.10
+
 ## 0.15.9
 
 ### Patch Changes