+++ packages/jazz-betterauth-client-plugin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # jazz-betterauth-client-plugin
 
+## 0.15.11
+
+### Patch Changes
+
+- jazz-betterauth-server-plugin@0.15.11
+
+## 0.15.10
+
+### Patch Changes
+
+- jazz-betterauth-server-plugin@0.15.10
+
 ## 0.15.9
 
 ### Patch Changes